Reorganizing, streamlining, updating comments
Restructuring this project to be more organized. This commit is not fully functional as it is as I have deleted structures in some scripts and not others, but still working on. In addition to edits like deleting unnecessary stuff, I have put all the scripts associated with each step of the process into separate folders.